refactor: FormTemplate -> ComplexAttributeTemplate
