+++ packages/dm-core-plugins/src/form/fields/ArrayField/templates/ArrayComplexTemplate.tsx
@@ -6,7 +6,7 @@ import AddObject from '../../../components/AddObjectButton'
 import { OpenObjectButton } from '../../../components/OpenObjectButton'
 import RemoveObject from '../../../components/RemoveObjectButton'
 import { useRegistryContext } from '../../../context/RegistryContext'
-import FormTemplate from '../../../templates/shared/FormTemplate'
+import { ComplexAttributeTemplate } from '../../../templates'
 import { TArrayTemplate } from '../../../types'
 import {
   getCanOpenOrExpand,
@@ -34,9 +34,9 @@ export const ArrayComplexTemplate = (props: TArrayTemplate) => {
     onOpen
   )
   return (
-    <FormTemplate>
-      <FormTemplate.Header>
-        <FormTemplate.Header.Title
+    <ComplexAttributeTemplate>
+      <ComplexAttributeTemplate.Header>
+        <ComplexAttributeTemplate.Header.Title
           canExpand={canExpand}
           objectIsNotEmpty={isDefined}
           canOpen={canOpen}
@@ -53,7 +53,7 @@ export const ArrayComplexTemplate = (props: TArrayTemplate) => {
           icon={list}
           uiAttribute={uiAttribute}
         />
-        <FormTemplate.Header.Actions uiAttribute={uiAttribute}>
+        <ComplexAttributeTemplate.Header.Actions uiAttribute={uiAttribute}>
           {canOpen && (
             <OpenObjectButton
               viewId={namePath}
@@ -79,9 +79,9 @@ export const ArrayComplexTemplate = (props: TArrayTemplate) => {
                 onAdd={() => setInitialValue([])}
               />
             ))}
-        </FormTemplate.Header.Actions>
-      </FormTemplate.Header>
-      <FormTemplate.Content
+        </ComplexAttributeTemplate.Header.Actions>
+      </ComplexAttributeTemplate.Header>
+      <ComplexAttributeTemplate.Content
         expanded={!!isExpanded}
         canExpand={!!canExpand}
         padding='px-2 pt-2'
@@ -91,7 +91,7 @@ export const ArrayComplexTemplate = (props: TArrayTemplate) => {
           onOpen={onOpen}
           viewConfig={getExpandViewConfig(uiAttribute)}
         />
-      </FormTemplate.Content>
-    </FormTemplate>
+      </ComplexAttributeTemplate.Content>
+    </ComplexAttributeTemplate>
   )
 }
