+++ packages/dm-core-plugins/src/form/templates/index.ts
@@ -1,0 +1,1 @@
+export * from './ComplexAttributeTemplate/ComplexAttributeTemplate'
